requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
protobuf==3.14.0
python-dotenv==0.15.0
six==1.15.0
Werkzeug==1.0.1
yapf==0.30.0
